requirements.txt
openai>=1.30.0
PyQt6>=6.7.0
tiktoken>=0.7.0
PyYAML>=6.0.1
pydantic>=2.7.0
